+++ src/postorius/tests/mailman_api_tests/__init__.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-# Copyright (C) 2012-2014 by the Free Software Foundation, Inc.
+# Copyright (C) 2012-2015 by the Free Software Foundation, Inc.
 #
 # This file is part of Postorius.
 #
@@ -14,34 +14,3 @@
 #
 # You should have received a copy of the GNU General Public License along with
 # Postorius.  If not, see <http://www.gnu.org/licenses/>.
-
-from mailmanclient.tests.utils import FakeMailmanClient
-
-from mock import patch
-from django.test import TestCase
-
-
-def setup_module():
-    FakeMailmanClient.setUp()
-
-def teardown_module():
-    FakeMailmanClient.tearDown()
-
-
-class MMTestCase(TestCase):
-
-    def _pre_setup(self):
-        super(MMTestCase, self)._pre_setup()
-        self.mm_client = FakeMailmanClient(
-            'http://localhost:8001/3.0', "restadmin", "restpass")
-        self.mm_client_patcher = patch('postorius.utils.Client', lambda *a, **kw: self.mm_client)
-        self.mm_client_patcher.start()
-        self.mm_client.create_domain(
-            'example.com',
-            contact_address='postmaster@example.com',
-            base_url='lists.example.com')
-
-    def _post_teardown(self):
-        self.mm_client_patcher.stop()
-        self.mm_client.delete_domain('example.com')
-        super(MMTestCase, self)._post_teardown()