+++ src/postorius/tests/mailman_api_tests/test_list_new.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-# Copyright (C) 2012-2014 by the Free Software Foundation, Inc.
+# Copyright (C) 2012-2015 by the Free Software Foundation, Inc.
 #
 # This file is part of Postorius.
 #
@@ -19,33 +19,55 @@ import logging
 
 from django.contrib.auth.models import User
 from django.core.urlresolvers import reverse
+from django.test import Client, SimpleTestCase
+from django.test.utils import override_settings
+from urllib2 import HTTPError
 
-from postorius.tests.mailman_api_tests import MMTestCase
+from postorius.utils import get_client
+from postorius.tests import MM_VCR
 
 
 logger = logging.getLogger(__name__)
+vcr_log = logging.getLogger('vcr')
+vcr_log.setLevel(logging.WARNING)
 
 
-class ListCreationTest(MMTestCase):
+API_CREDENTIALS = {'MAILMAN_API_URL': 'http://localhost:9001',
+                   'MAILMAN_USER': 'restadmin',
+                   'MAILMAN_PASS': 'restpass'}
+
+
+@override_settings(**API_CREDENTIALS)
+class ListCreationTest(SimpleTestCase):
     """Tests for the new list page."""
 
+    @MM_VCR.use_cassette('test_list_creation.yaml')
     def setUp(self):
+        self.client = Client()
         self.user = User.objects.create_user('user', 'user@example.com', 'pwd')
         self.superuser = User.objects.create_superuser('su', 'su@example.com',
                                                        'pwd')
+        try:
+            self.domain = get_client().create_domain('example.com')
+        except HTTPError:
+            self.domain = get_client().get_domain('example.com')
 
+    @MM_VCR.use_cassette('test_list_creation.yaml')
     def tearDown(self):
         self.user.delete()
         self.superuser.delete()
+        for mlist in get_client().lists:
+            mlist.delete()
 
     def test_permission_denied(self):
         self.client.login(username='user', password='pwd')
         response = self.client.get(reverse('list_new'))
-        self.assertRedirects(
-            response,
-            '/postorius/accounts/login/?next=/postorius/lists/new/')
+        self.assertEqual(
+            response['location'],
+            'http://testserver/postorius/accounts/login/?next=/lists/new/')
 
-    def test_new_list_created(self):
+    @MM_VCR.use_cassette('test_list_creation.yaml')
+    def test_new_list_created_with_owner(self):
         self.client.login(username='su', password='pwd')
         post_data = {'listname': 'a_new_list',
                      'mail_host': 'example.com',
@@ -53,6 +75,6 @@ class ListCreationTest(MMTestCase):
                      'advertised': 'True',
                      'description': 'A new list.'}
         self.client.post(reverse('list_new'), post_data)
-        a_new_list = self.mm_client.get_list('a_new_list@example.com')
+        a_new_list = get_client().get_list('a_new_list@example.com')
         self.assertEqual(a_new_list.fqdn_listname, u'a_new_list@example.com')
         self.assertEqual(a_new_list.owners, [u'owner@example.com'])