+++ testing/__init__.py
@@ -1,5 +1,5 @@
-#-*- coding: utf-8 -*-
-# Copyright (C) 1998-2015 by the Free Software Foundation, Inc.
+# -*- coding: utf-8 -*-
+# Copyright (C) 2012-2014 by the Free Software Foundation, Inc.
 #
 # This file is part of Postorius.
 #
@@ -7,7 +7,6 @@
 # the terms of the GNU General Public License as published by the Free
 # Software Foundation, either version 3 of the License, or (at your option)
 # any later version.
-#
 # Postorius is distributed in the hope that it will be useful, but WITHOUT
 # ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
 # FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License for
@@ -16,5 +15,16 @@
 # You should have received a copy of the GNU General Public License along with
 # Postorius.  If not, see <http://www.gnu.org/licenses/>.
 
-"""This is just an empty settings file to make Sphinx able to export
-the DJANGO_SETTINGS_MODULE env variable before building the docs."""
+import os
+import vcr
+
+from django.conf import settings
+
+
+TEST_DIR = os.path.abspath(os.path.dirname(__file__))
+FIXTURES_DIR = os.path.join(TEST_DIR, 'fixtures', 'vcr_cassettes')
+
+
+MM_VCR = vcr.VCR(serializer='json',
+                 cassette_library_dir=FIXTURES_DIR,
+                 match_on=['uri', 'method'])